Rename file to match renamed class
